+++ cxroots/iterative_methods.py
@@ -1,5 +1,7 @@
 import logging
 
+from mpmath import mp, mpmathify
+from mpmath.calculus.optimization import Muller
 from numpy import inf
 
 
@@ -10,7 +12,7 @@ def iterate_to_root(
     step_tol=1e-12,
     root_tol=1e-12,
     max_iter=20,
-    attempt_best=False,
+    refine_roots_beyond_tol=False,
     callback=None,
 ):
     """
@@ -30,18 +32,17 @@ def iterate_to_root(
         The derivative of f.
     step_tol: float, optional
         The routine ends if the step size, dx, between sucessive
-        iterations satisfies abs(dx) < step_tol and attempt_best is False.
+        iterations satisfies abs(dx) < step_tol and refine_roots_beyond_tol is False.
     root_tol: float, optional
-        The routine ends if abs(f(x)) < root_tol and attempt_best is False.
+        A root, x, is only returned if abs(f(x)) < root_tol, otherwise None is returned
     max_iter : int, optional
         The routine ends after max_iter iterations.
-    attempt_best : bool, optional
-        If True then routine ends if the error of the previous iteration,
-        x0, was at least as good as the current iteration, x, in the
-        sense that abs(f(x)) >= abs(f(x0)) and the previous iteration
-        satisfied either abs(dx0) < step_tol or abs(f(x0)) < root_tol.  In
-        this case the previous iteration is returned as the approximation
-        of the root.
+    refine_roots_beyond_tol : bool, optional
+        If True then the routine ends only once the error of the previous iteration,
+        x0, was at least as good as the current iteration, x, in the sense that
+        abs(f(x)) >= abs(f(x0)), and the previous iteration satisfied
+        abs(dx0) < step_tol. In this case the previous iteration is returned as the
+        approximation of the root, provided that it satisfies abs(f(x)) < root_tol
     callback : function, optional
         After each iteration callback(x, dx, f(x), iteration) will be
         called where 'x' is the current iteration of the estimated root,
@@ -52,7 +53,7 @@ def iterate_to_root(
     Returns
     -------
     complex
-        An approximation for a root of f.  If the rootfinding was
+        An approximation for a root of f. If the rootfinding was
         unsucessful then None will be returned instead.
     """
     logger = logging.getLogger(__name__)
@@ -60,7 +61,9 @@ def iterate_to_root(
 
     if df is not None:
         try:
-            root, err = newton(x0, f, df, step_tol, 0, max_iter, attempt_best, callback)
+            root, err = newton(
+                x0, f, df, step_tol, 0, max_iter, refine_roots_beyond_tol, callback
+            )
         except (RuntimeError, OverflowError):
             return None
     else:
@@ -70,7 +73,15 @@ def iterate_to_root(
 
         x1, x2, x3 = x0, x0 * (1 + 1e-8) + 1e-8j, x0 * (1 - 1e-8) - 1e-8j
         root, err = muller(
-            x1, x2, x3, f_muller, step_tol, 0, max_iter, attempt_best, callback
+            x1,
+            x2,
+            x3,
+            f_muller,
+            step_tol,
+            0,
+            max_iter,
+            refine_roots_beyond_tol,
+            callback,
         )
 
     if err < root_tol:
@@ -83,9 +94,9 @@ def muller(
     x3,
     f,
     step_tol=1e-12,
-    root_tol=1e-12,
+    root_tol=0,
     max_iter=20,
-    attempt_best=False,
+    refine_roots_beyond_tol=False,
     callback=None,
 ):
     """
@@ -105,12 +116,12 @@ def muller(
         Function of a single variable which we seek to find a root of.
     step_tol: float, optional
         The routine ends if the step size, dx, between sucessive
-        iterations satisfies abs(dx) < step_tol and attempt_best is False.
+        iterations satisfies abs(dx) < step_tol and refine_roots_beyond_tol is False.
     root_tol: float, optional
-        The routine ends if abs(f(x)) < root_tol and attempt_best is False.
+        The routine ends if abs(f(x)) < root_tol and refine_roots_beyond_tol is False.
     max_iter : int, optional
         The routine ends after max_iter iterations.
-    attempt_best : bool, optional
+    refine_roots_beyond_tol : bool, optional
         If True then routine ends if the error of the previous iteration,
         x0, was at least as good as the current iteration, x, in the
         sense that abs(f(x)) >= abs(f(x0)) and the previous iteration
@@ -131,9 +142,6 @@ def muller(
     float
         abs(f(x)) where x is the final approximation for the root of f.
     """
-    from mpmath import mp, mpmathify
-    from mpmath.calculus.optimization import Muller
-
     logger = logging.getLogger(__name__)
 
     # mpmath insists on functions accepting mpc
@@ -141,14 +149,14 @@ def muller(
         return mpmathify(f(complex(z)))
 
     mull = Muller(mp, f_mpmath, (x1, x2, x3), verbose=False)
-    iteration = 0
     x0 = x3
 
     x, err = x0, abs(f(x0))
     err0, dx0 = inf, inf
     try:
-        for x, dx in mull:
-            err = abs(f_mpmath(x))
+        for iteration, (x, dx) in enumerate(mull):
+            y = f_mpmath(x)
+            err = abs(y)
             logger.debug(
                 str(iteration)
                 + " x="
@@ -159,18 +167,18 @@ def muller(
                 + str(dx)
             )
 
-            if callback is not None and callback(x, dx, err, iteration + 1):
+            if callback is not None and callback(x, dx, y, iteration + 1):
                 break
 
             if (
-                not attempt_best
+                not refine_roots_beyond_tol
                 and (abs(dx) < step_tol or err < root_tol)
                 or iteration > max_iter
             ):
                 break
 
             if (
-                attempt_best
+                refine_roots_beyond_tol
                 and (abs(dx0) < step_tol or err0 < root_tol)
                 and err >= err0
             ):
@@ -179,10 +187,9 @@ def muller(
                 x, err = x0, err0
                 break
 
-            iteration += 1
             x0 = x
 
-            if attempt_best:
+            if refine_roots_beyond_tol:
                 # record previous error for comparison
                 dx0, err0 = dx, err
 
@@ -202,9 +209,9 @@ def newton(
     f,
     df,
     step_tol=1e-12,
-    root_tol=1e-12,
+    root_tol=0,
     max_iter=20,
-    attempt_best=False,
+    refine_roots_beyond_tol=False,
     callback=None,
 ):
     """
@@ -224,12 +231,12 @@ def newton(
         derivative of the function f(x) at the point x
     step_tol: float, optional
         The routine ends if the step size, dx, between sucessive
-        iterations satisfies abs(dx) < step_tol and attempt_best is False.
+        iterations satisfies abs(dx) < step_tol and refine_roots_beyond_tol is False.
     root_tol: float, optional
-        The routine ends if abs(f(x)) < root_tol and attempt_best is False.
+        The routine ends if abs(f(x)) < root_tol and refine_roots_beyond_tol is False.
     max_iter : int, optional
         The routine ends after max_iter iterations.
-    attempt_best : bool, optional
+    refine_roots_beyond_tol : bool, optional
         If True then routine ends if the error of the previous iteration,
         x0, was at least as good as the current iteration, x, in the
         sense that abs(f(x)) >= abs(f(x0)) and the previous iteration
@@ -263,17 +270,17 @@ def newton(
         if callback is not None and callback(x, dx, y, iteration + 1):
             break
 
-        if not attempt_best and (abs(dx) < step_tol or abs(y) < root_tol):
+        if not refine_roots_beyond_tol and (abs(dx) < step_tol or abs(y) < root_tol):
             break
 
         if (
-            attempt_best
+            refine_roots_beyond_tol
             and (abs(dx0) < step_tol or abs(y0) < root_tol)
-            and abs(y) > abs(y0)
+            and abs(y) >= abs(y0)
         ):
             break
 
-        if attempt_best:
+        if refine_roots_beyond_tol:
             # store previous dx and y
             dx0, y0 = dx, y
 
@@ -281,7 +288,7 @@ def newton(
     return x, abs(y)
 
 
-def secant(x1, x2, f, step_tol=1e-12, root_tol=1e-12, max_iter=30, callback=None):
+def secant(x1, x2, f, step_tol=1e-12, root_tol=0, max_iter=30, callback=None):
     """
     Find an approximation to a point xf such that f(xf)=0 for a
     scalar function f using the secant method.  The method requires
@@ -300,9 +307,9 @@ def secant(x1, x2, f, step_tol=1e-12, root_tol=1e-12, max_iter=30, callback=None
         Function of a single variable which we seek to find a root of.
     step_tol: float, optional
         The routine ends if the step size, dx, between sucessive
-        iterations satisfies abs(dx) < step_tol and attempt_best is False.
+        iterations satisfies abs(dx) < step_tol and refine_roots_beyond_tol is False.
     root_tol: float, optional
-        The routine ends if abs(f(x)) < root_tol and attempt_best is False.
+        The routine ends if abs(f(x)) < root_tol and refine_roots_beyond_tol is False.
     max_iter : int, optional
         The routine ends after max_iter iterations.
     callback : function, optional